requirements.txt
tensorflow==2.6.2
tensorflow_io==0.21.0
tensorflow_probability==0.14.1
tensorflow_addons==0.15.0

wandb==0.12.9
loguru==0.6.0

pandas==1.3.5
tqdm==4.62.3
matplotlib==3.5.1
seaborn==0.11.2
PyYAML>=5.1
scikit-learn>=1.0.0
librosa==0.9.1

efficientnet==1.1.1
keras-cv-attention-models==1.2.12
git+https://github.com/sebastian-sz/resnet-rs-keras@main
timm==0.5.4
tfimm==0.2.5